+++ tests/com/yanncebron/m68kplugin/parser/MnemonicGeneratedParserDataTest.java
@@ -81,7 +81,7 @@ public class MnemonicGeneratedParserDataTest extends M68kParsingTestCase {
 
     dump("* Instructions count: " + M68kTokenGroups.INSTRUCTIONS.getTypes().length);
 
-    assertEquals(5495, total);
+    assertEquals(5916, total);
     assertEmpty(failedVariants);
     assertEquals(781, deprecated);
   }